HSEARCH-3398 Move PojoWorkPlan to SPI (for real this time...)
In ab07d7d5fd6652b7e92e3adfd9b0820980feb631 I did everything necessary
to move PojoWorkPlan to SPI (i.e. I made sure to wrap it with APIs where
needed), but I forgot to move the interface itself to SPI...
