+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/impl/InTransactionWorkQueueSynchronization.java
@@ -13,7 +13,7 @@ import java.util.concurrent.CompletableFuture;
 import javax.transaction.Synchronization;
 
 import org.hibernate.search.mapper.orm.logging.impl.Log;
-import org.hibernate.search.mapper.pojo.mapping.PojoWorkPlan;
+import org.hibernate.search.mapper.pojo.work.spi.PojoWorkPlan;
 import org.hibernate.search.util.impl.common.LoggerFactory;
 
 /**
