+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/session/spi/PojoSearchManagerDelegate.java
@@ -8,7 +8,7 @@ package org.hibernate.search.mapper.pojo.session.spi;
 
 import java.util.Collection;
 
-import org.hibernate.search.mapper.pojo.mapping.PojoWorkPlan;
+import org.hibernate.search.mapper.pojo.work.spi.PojoWorkPlan;
 import org.hibernate.search.mapper.pojo.search.spi.PojoSearchTargetDelegate;
